Enable evaluation of netlists synthesized for Xilinx FPGAs.

+++ inc/Hardware/FaultSet.hpp
@@ -11,12 +11,13 @@
 
 #pragma once
 
-#include "Hardware/Fault.hpp"
 #include <cstdint>
 #include <vector>
 
+#include "Hardware/Fault.hpp"
+
 class FaultSet {
-public:
+ public:
   /**
    * @brief Constructor of a FaultSet.
    *
@@ -31,7 +32,7 @@ public:
    *
    * @return The number of Faults in the FaultSet.
    */
-  size_t GetNumberOfFaultsInSet() const;
+  uint64_t GetNumberOfFaultsInSet() const;
 
   /**
    * @brief Accessor of the number of simulations for which the FaultSet was
@@ -40,7 +41,7 @@ public:
    * @return Returns the number of simulations for which the FaultSet was
    * effective.
    */
-  size_t GetNumberOfEffectiveSimulations() const;
+  uint64_t GetNumberOfEffectiveSimulations() const;
 
   /**
    * @brief Accessor function for a Fault of this FaultSet.
@@ -48,7 +49,7 @@ public:
    * @param index The index of the Fault, which is accessed.
    * @return Returns the Fault which is accessed.
    */
-  Fault const *GetFault(size_t index) const;
+  Fault const *GetFault(uint64_t index) const;
 
   /**
    * @brief Increments the number of simulations for which the FaultSet was
@@ -63,16 +64,16 @@ public:
    * @param number The number by which the counter of effective simulations
    * should be increased.
    */
-  void FaultSetWasEffective(size_t number);
+  void FaultSetWasEffective(uint64_t number);
 
-  void TryToInduceFaults(uint64_t& value, uint64_t signal_index, uint64_t clock_cycle);
+  void TryToInduceFaults(uint64_t &value, uint64_t signal_index,
+                         uint64_t clock_cycle);
 
-private:
+ private:
   const std::vector<Fault const *> faults_;
-  size_t number_of_effective_simulations_{0};
+  uint64_t number_of_effective_simulations_{0};
 
   // FIXME: Add a variable input_causing_effective_faults,
   // which stores which input causes a faultset to be active.
   // Either global input or input to faulted gates,
 };
-
