+++ src/Hardware/FaultSet.cpp
@@ -10,11 +10,13 @@
 
 #include "Hardware/FaultSet.hpp"
 
-FaultSet::FaultSet(std::vector<Fault const *> &faults) : faults_(faults){};
+FaultSet::FaultSet(std::vector<Fault const *> &faults) : faults_(faults) {};
 
-size_t FaultSet::GetNumberOfFaultsInSet() const { return this->faults_.size(); }
+uint64_t FaultSet::GetNumberOfFaultsInSet() const {
+  return this->faults_.size();
+}
 
-size_t FaultSet::GetNumberOfEffectiveSimulations() const {
+uint64_t FaultSet::GetNumberOfEffectiveSimulations() const {
   return this->number_of_effective_simulations_;
 }
 
@@ -22,21 +24,19 @@ void FaultSet::FaultSetWasEffective() {
   ++this->number_of_effective_simulations_;
 }
 
-void FaultSet::FaultSetWasEffective(size_t number) {
+void FaultSet::FaultSetWasEffective(uint64_t number) {
   this->number_of_effective_simulations_ += number;
 }
 
-Fault const *FaultSet::GetFault(size_t index) const {
+Fault const *FaultSet::GetFault(uint64_t index) const {
   return this->faults_[index];
 };
 
-void FaultSet::TryToInduceFaults(uint64_t& value, uint64_t signal_index, uint64_t clock_cycle) {
-  for (const Fault* fault : faults_) {
+void FaultSet::TryToInduceFaults(uint64_t &value, uint64_t signal_index,
+                                 uint64_t clock_cycle) {
+  for (const Fault *fault : faults_) {
     if (fault->IsSignalFaulted(signal_index, clock_cycle)) {
       value = fault->ComputeFaultEffect(value);
     }
   }
 }
-
-
-
